- Distributed version ported and tested on Win7 64-bit. - little fix on the allocator.hpp, freesegment now release aligned memory   to avoid stack corruption on Win platform that requires symmetric   primitives for aligned memory allocation and deallocation. - Winsock.h (windows.h) and winsock2.h incompatbility issue partially solved.   <ff/dnode.h> should be included before other includes to avoid the problem.   It includes winsock2.h that is required by zeromq.

+++ ff/mapper.hpp
@@ -76,8 +76,8 @@ public:
  
         mask = size-1;
         
-        for(size_t i=0;i<num_cores;++i) CList.push_back(i);
-        for(size_t i=num_cores,j=0; i<size;++i,j++) CList.push_back(j);
+        for(unsigned int i=0;i<num_cores;++i) CList.push_back(i);
+        for(unsigned int i=num_cores,j=0; i<size;++i,j++) CList.push_back(j);
         
         rrcnt=0;
     }
@@ -114,7 +114,7 @@ public:
             if (*_str == '\0')	break;
         } while (1);
 
-        unsigned long size=List.size();
+        size_t size=List.size();
         if (!isPowerOf2(size)) {
             size=nextPowerOf2(size);
             List.reserve(size);
